Include unit test configuration
Get the project with basic setup to unit test.

Modifies the following:

tslint - To accept jest keywords;
babelrc - Moves it to js file version;
package - Includes all the necessary dependencies

Also, this modifies the typescript files to export
interfaces and declared types on sight, as when
exporting afterwards causes errors with coverage.